Move _optional.py out of the compat folder (until I figure out submodules).
